refactor: Split `createApplicationCommand` into two files, for guild and global commands respectively.
